+++ .eslintrc.json
@@ -19,6 +19,7 @@
               "endOfLine": "auto"
             }
           ],
-        "no-console":"off"
+        "no-console":"off",
+        "no-unused-vars": "warn"
     }
 }
