+++ app/app.js
@@ -1,6 +1,4 @@
 const express = require("express");
-const cors = require("cors");
-const { cacheControl } = require("./middlewares/middlewares");
 const router = require("./routes/index");
 
 const app = express();
