+++ app/middlewares/middlewares.js
@@ -1,5 +1,5 @@
 const checkMimeType = require("./check-mime-type");
-const isUpload = require("./has-file");
+const hasFile = require("./has-file");
 const mediaError = require("./media-error-handler");
 const cacheControl = require("./cache-control");
 const basicAuth = require("./basic-auth");
@@ -7,7 +7,7 @@ const authErrorHandler = require("./auth-error-handler");
 
 module.exports = {
   checkMimeType,
-  isUpload,
+  hasFile,
   mediaError,
   cacheControl,
   basicAuth,
