+++ app/routes/secret.js
@@ -1,14 +1,9 @@
 const router = require("express").Router();
 const { basicAuth, authErrorHandler } = require("../middlewares/middlewares");
+const sendSecret = require("../controllers/secret");
 
 router.use("/", basicAuth);
 router.use("/", authErrorHandler);
 
-router.get("/", async (req, res) => {
-  try {
-    res.status(200).send("You have found many, many secrets");
-  } catch (err) {
-    console.log(err.message);
-  }
-});
+router.get("/", sendSecret);
 module.exports = router;
